#17 アニメーションを別ファイルで定義
