+++ my-app/src/components/HeaderComponent/Header.tsx
@@ -4,35 +4,7 @@ import HeaderMenu from "./HeaderMenu";
 import HeaderTitle from "./HeaderTitle";
 import { FiMenu } from "react-icons/fi";
 import { MdClose } from "react-icons/md";
-
-const IconsAnimationVariants = {
-  hidden: { opacity: 0 },
-  visible: { opacity: 1, rotate: 360 },
-  exit: { opacity: 0 },
-};
-
-const IconsAnimation = {
-  variants: IconsAnimationVariants,
-  initial: "hidden",
-  animate: "visible",
-  exit: "exit",
-  whileHover: { scale: 1.2 },
-  whileTap: { scale: 0.9 },
-  transition: { duration: 0.1 },
-};
-
-const MenuAnimationVariants = {
-  hidden: { opacity: 0, y: -20 },
-  visible: { opacity: 1, y: 0 },
-  exit: { opacity: 0, y: -20 },
-};
-
-const MenuAnimation = {
-  variants: MenuAnimationVariants,
-  initial: "hidden",
-  animate: "visible",
-  exit: "exit",
-};
+import { IconsAnimation, MenuAnimation } from "../Animations/Animations";
 
 const Header = () => {
   const [MenuOpen, setMenuOpen] = useState(false);
